pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kion-mcp-server"
version = "0.1.0"
description = "Kion MCP Server - Model Context Protocol server for Kion API integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.10.5",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
kion-mcp-server = "kion_mcp.server:run"

[tool.hatch.build.targets.wheel]
packages = ["src/kion_mcp"]
